% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_convert_CEP_into_city_state}
\alias{CS_convert_CEP_into_city_state}
\title{Switch CEP into city and state}
\usage{
CS_convert_CEP_into_city_state(cep_string)
}
\arguments{
\item{vec}{A vector CEPs extracted from}
}
\value{
A named vector with city and state
}
\description{
Convert a CEP into a vector with city and state through a check in cepdarua.net
}
\examples{
CS_convert_CEP_into_city_state(CS_get_bus_CEP(df$Endereco))

}
