% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_convert_mon_to_num}
\alias{CS_convert_mon_to_num}
\title{Convert Monetary values to numeric}
\usage{
CS_convert_mon_to_num(vec)
}
\arguments{
\item{vec}{A vector with businesses CNAEs, preferably collected from CS_get_bus_df}
}
\value{
A named vector with CNAEs sector and character values indicating `yes` or `no`
}
\description{
Convert the business CNAE to our dictionary (check `cnae_vec` object)
}
\examples{
CS_CNAE_conv(CS_get_CNAE(df$Atividade.economica))

}
