% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_convert_state_to_region}
\alias{CS_convert_state_to_region}
\title{Convert states into regions}
\usage{
CS_convert_state_to_region(vec)
}
\arguments{
\item{vec}{A vector with businesses states, preferably collected from CS_get_bus_states}
}
\value{
A character vector containing these businesses region
}
\description{
Changes states into regions through our package internal data
}
\examples{
CS_get_total_ksoc(unique(df$Capital.social))

}
