% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_get_big_n_employees}
\alias{CS_get_big_n_employees}
\title{Get biggest income from businesses}
\usage{
CS_get_big_n_employees(vec)
}
\arguments{
\item{vec}{A vector with businesses incomes preferably collected from CS_get_bus_df}
}
\value{
An integer, following our dictionary
}
\description{
Filter the biggest income from a business owner
}
\examples{
CS_get_big_n_employees(df$Número.de.funcionários)

}
