% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\encoding{UTF-8}
\name{CS_get_big_size}
\alias{CS_get_big_size}
\title{Get biggest business size from businesses}
\usage{
CS_get_big_size(vec)
}
\arguments{
\item{vec}{A vector with businesses sizes preferably collected from CS_get_bus_df}
}
\value{
An integer, following our dictionary
}
\description{
Filter the biggest business from a business owner
}
\examples{
CS_get_big_size(df$Número.de.funcionários)
}
