% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_get_bus_CEP}
\alias{CS_get_bus_CEP}
\title{Extract CEP from address}
\usage{
CS_get_bus_CEP(vec)
}
\arguments{
\item{vec}{A vector with businesses adresses, preferably collected from CS_get_bus_states}
}
\value{
A character vector containing these businesses CEPs
}
\description{
Extract CEP from an address displayed
}
\examples{
CS_get_bus_CEP(df$Endereço)

}
