% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{CS_get_bus_df}
\alias{CS_get_bus_df}
\title{Get businesses data.frame}
\usage{
CS_get_bus_df(path, cpf = NULL)
}
\arguments{
\item{path}{A path that leads to a Consultasocio page}

\item{cpf}{A character vector containing the cpf numbers available in the Consultasocio page}
}
\value{
A business data.frame with many variables, basically following all information available in their page
}
\description{
Creates a business data.frame from an owner page
}
\details{
It's a scraping function written with rvest and returns a basically all information available in that page, which demands some processing later.
To use our functions, it's advised to use Web Archive to save and backup that data. With that method, it's possible to follow a longitudinal study.
It works both with the direct links or with a downloaded html page.
In case of more than one page, we recommend lapply and bind_rows functions to tie these different pages.
}
\examples{
CS_get_bus_df("https://web.archive.org/web/20211221092935/https://www.consultasocio.com/q/sa/arthur-moledo-do-val")
do.call(bind_rows, lapply(X = c("https://web.archive.org/web/20220607001536/https://www.consultasocio.com/q/sa/paulo-antonio-skaf",
                                "https://web.archive.org/web/20220607002216/https://www.consultasocio.com/q/sa/paulo-antonio-skaf?page=2"
                                ),
                          FUN = CS_get_bus_df,
                          cpf = "083628"
                          )
        )

}
