% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_get_CNAE}
\alias{CS_get_CNAE}
\title{Get CNAE codes}
\usage{
CS_get_CNAE(vec)
}
\arguments{
\item{vec}{A vector with businesses CNAEs, preferably collected from CS_get_bus_df}
}
\value{
A character vector with its businesses CNAEs
}
\description{
Extract CNAE codes from businesses
}
\examples{
CS_get_CNAE(df$Atividade.economica)

}
