% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_df.R
\name{CS_get_summary_df}
\alias{CS_get_summary_df}
\title{Tie all summary vec in a data.frame}
\usage{
CS_get_summary_df(vec_list)
}
\arguments{
\item{vec_list}{A list containing multiple summary vectors}
}
\value{
A formatted summary data.frame
}
\description{
Tie your vector into a data.frame and organize them to use them as a factor
}
