% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_df.R
\encoding{UTF-8}
\name{CS_get_summary_vec_reg}
\alias{CS_get_summary_vec_reg}
\title{Creates a summary data.frame from business data.frame}
\usage{
CS_get_summary_vec_reg(df)
}
\arguments{
\item{df}{A business owner page data.frame}
}
\value{
A named character vector with a summary data
}
\description{
Builds a summary vector from a business owner data.frame, to be bound with many other business owners data
}
\examples{
CS_get_summary_vec_reg(Skaf_bus_df)
}
