% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling.R
\name{CS_get_total_ksoc}
\alias{CS_get_total_ksoc}
\title{Collect and sum Share Capital}
\usage{
CS_get_total_ksoc(vec)
}
\arguments{
\item{vec}{A vector with businesses share capital, preferably collected from CS_get_bus_df}
}
\value{
A numeric element
}
\description{
Process values, retire NAs and sum them up
}
\examples{
CS_get_total_ksoc(unique(df$Capital.social))

}
