% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_df.R
\name{CS_read_board_csv}
\alias{CS_read_board_csv}
\title{Read boards csv}
\usage{
CS_read_board_csv(csv_path)
}
\arguments{
\item{csv_path}{A path that leads to a .csv file containing businessman links}
}
\value{
A data.frame with the business owners summary data, organized with default values provided by Consultasocio
}
\description{
Scrape, wrangle and organizes data from business owners pages
}
\details{
Reads a complete csv file, following the patterns shown in our vignette, aiding us in our replicability and organization.
There are basically four columns: owner name, organization, cpf and links.
It already automatizes some processes, like CPF selection and multiple pages.
}
\examples{
CS_read_board_csv("diretores.csv")

}
