% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_df.R
\name{CS_read_files_backup}
\alias{CS_read_files_backup}
\title{Read backup files}
\usage{
CS_read_files_backup(ent_name, period)
}
\arguments{
\item{ent_name}{Your organization to research, organized in your folder}

\item{period}{Period of activities of your board}
}
\value{
A data.frame with the business owners summary data, organized with default values provided by Consultasocio
}
\description{
Scrape, wrangle and organizes data from business owners pages backup in your project
}
\details{
Reads your backup, following the patterns shown in our vignette, aiding us in our replicability and organization.
It already automatizes some processes, like CPF selection and multiple pages.
Should have a board csv with CPF data and Consultasocio links and proper directories organization.
For more details, check our vignette.
}
\examples{
CS_read_files_backup("FIESC", "21-24")

}
