% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{WArch_store_page}
\alias{WArch_store_page}
\title{Save Consultasocio links}
\usage{
WArch_store_page(link, WArch_auth)
}
\arguments{
\item{link}{A link to a consultasocio page that you want to save}

\item{WArch_aut}{A string containing your Authorization data. To see how to write it, check our example}
}
\value{
A character string containing a web.archive link
}
\description{
Archive your links directly with SPN2 API
}
\details{
Still in development, we're trying to improve it. It is basically a code forcing to store your link, so it could take a while.
To use that function, you're going to need to create an account in web.archive.org and then request a key for their API.
For a better introduction to how to use this API, check their [documentation](https://docs.google.com/document/d/1Nsv52MvSjbLb2PCpHlat0gkzw0EvtSgpKHu4mk0MnrA/view).
}
\examples{
CS_WArch_store_page("https://www.consultasocio.com/q/sa/arthur-moledo-do-val", "YOUR_KEY <YOUR_SECRET>")

}
